refactor: move searchdate models into models.cs
